+++ packages/camera_avfoundation/ios/camera_extended_ios/Sources/camera_avfoundation/Camera.swift
@@ -63,6 +63,10 @@ protocol Camera: FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate,
 
   func captureToFile(completion: @escaping (_ path: String?, _ error: FlutterError?) -> Void)
 
+  func captureToMemory(
+    completion: @escaping (_ data: FlutterStandardTypedData?, _ error: FlutterError?) -> Void
+  )
+
   func lockCaptureOrientation(_ orientation: FCPPlatformDeviceOrientation)
   func unlockCaptureOrientation()